test(demoqa): create Helpers.ElementDisplayInspector to refactoring "BtnTest" code
