+++ src/test/java/DemoQA/ElementsTests/BtnTest.java
@@ -1,12 +1,11 @@
 package DemoQA.ElementsTests;
 
+import DemoQA.Helpers.ElementDisplayInspector;
 import DemoQA.Pages.ButtonsPage;
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
@@ -16,9 +15,10 @@ import org.testng.annotations.Test;
 import java.time.Duration;
 
 public class BtnTest {
-    WebDriver driver;
     WebDriverWait wait;
     ButtonsPage buttonsPage;
+    ElementDisplayInspector inspector = new ElementDisplayInspector();
+    private WebDriver driver;
 
     @BeforeTest
     public void setUp() {
@@ -26,6 +26,7 @@ public class BtnTest {
         driver = new ChromeDriver(); // Example for Chrome
         wait = new WebDriverWait(driver, Duration.ofSeconds(10));
         driver.get("https://demoqa.com/buttons");
+        inspector.setDriver(driver);
         buttonsPage = new ButtonsPage();
     }
 
@@ -35,10 +36,10 @@ public class BtnTest {
 
         Actions action = new Actions(driver);
 
-        Assert.assertTrue(elementIsDisappear(buttonsPage.doubleClickMessageLocator));
+        Assert.assertTrue(inspector.elementIsDisappear(buttonsPage.doubleClickMessageLocator));
 
         action.doubleClick(doubleClickBtn).perform();
-        Assert.assertTrue(elementExists(buttonsPage.doubleClickMessageLocator));
+        Assert.assertTrue(inspector.elementExists(buttonsPage.doubleClickMessageLocator));
         WebElement message = driver.findElement(buttonsPage.doubleClickMessageLocator);
 
         assert message.getText().equals("You have done a double click");
@@ -50,10 +51,10 @@ public class BtnTest {
 
         Actions action = new Actions(driver);
 
-        Assert.assertTrue(elementIsDisappear(buttonsPage.rightClickMessageLocator));
+        Assert.assertTrue(inspector.elementIsDisappear(buttonsPage.rightClickMessageLocator));
 
         action.contextClick(rightClickBtn).perform();
-        Assert.assertTrue(elementExists(buttonsPage.rightClickMessageLocator));
+        Assert.assertTrue(inspector.elementExists(buttonsPage.rightClickMessageLocator));
         WebElement message = driver.findElement(buttonsPage.rightClickMessageLocator);
 
         assert message.getText().equals("You have done a right click");
@@ -64,41 +65,15 @@ public class BtnTest {
         WebElement clickMeBtn = driver.findElement(buttonsPage.clickMeBtnLocator);
         Actions action = new Actions(driver);
 
-        Assert.assertTrue(elementIsDisappear(buttonsPage.dynamicClickMessage));
+        Assert.assertTrue(inspector.elementIsDisappear(buttonsPage.dynamicClickMessage));
 
         action.click(clickMeBtn).perform();
-        Assert.assertTrue(elementExists(buttonsPage.dynamicClickMessage));
+        Assert.assertTrue(inspector.elementExists(buttonsPage.dynamicClickMessage));
         WebElement message = driver.findElement(buttonsPage.dynamicClickMessage);
 
         assert message.getText().equals("You have done a dynamic click");
     }
 
-    public Boolean elementExists(By locator) {
-        try {
-            System.out.println("Element is exist: " + locator.toString());
-            // Wait until the element is visible
-            WebElement button = wait.until(ExpectedConditions.visibilityOfElementLocated(locator));
-            // Perform actions on the element
-            System.out.println("end");
-            return button.isDisplayed();
-        } catch (Exception e) {
-            System.out.println("exeption");
-            return false;
-        }
-    }
-
-    public Boolean elementIsDisappear(By locator) {
-        try {
-            System.out.println("Element is disappear: " + locator.toString());
-            WebElement button = driver.findElement(locator);
-            // Perform actions on the element
-            System.out.println("end");
-            return button.isDisplayed();
-        } catch (Exception e) {
-            System.out.println("exeption");
-            return true;
-        }
-    }
 
     @AfterTest
     public void tearDown() {
